feat: Split settings into a nav menu with per-section pages
- Replace the single settings screen with a menu that pushes to sub-pages
- Add SettingsMenuRow (icon + label + chevron) for the menu
- Move Appearance, Notifications, Preferences, and About into their own routes
